+++ annlib/cost_function.h
@@ -13,7 +13,7 @@ namespace annlib
 	public:
 		virtual ~cost_function() = default;
 
-		virtual double calculate_costs(const mat_arr& net_output_rv,
+		virtual float calculate_costs(const mat_arr& net_output_rv,
 		                               const mat_arr& solution_rv) const = 0;
 
 		virtual void calculate_gradient(const mat_arr& net_output_rv,
@@ -22,13 +22,13 @@ namespace annlib
 
 		virtual void calculate_output_layer_error(const mat_arr& net_output_rv, const mat_arr& solution_rv,
 		                                          const mat_arr& output_layer_weighted_input_rv,
-		                                          const std::function<double(double)>& derivative_activation_function,
+		                                          const std::function<float(float)>& derivative_activation_function,
 		                                          mat_arr* output_layer_error_rv) const;
 	};
 
 	class quadratic_costs : public cost_function
 	{
-		double calculate_costs(const mat_arr& net_output_rv,
+		float calculate_costs(const mat_arr& net_output_rv,
 		                       const mat_arr& solution_rv) const override;
 
 		void calculate_gradient(const mat_arr& net_output_rv,
@@ -38,7 +38,7 @@ namespace annlib
 
 	class cross_entropy_costs : public cost_function
 	{
-		double calculate_costs(const mat_arr& net_output_rv,
+		float calculate_costs(const mat_arr& net_output_rv,
 		                       const mat_arr& solution_rv) const override;
 
 		void calculate_gradient(const mat_arr& net_output_rv,
@@ -48,7 +48,7 @@ namespace annlib
 		void calculate_output_layer_error(const mat_arr& net_output_rv,
 		                                  const mat_arr& solution_rv,
 		                                  const mat_arr& output_layer_weighted_input_rv,
-		                                  const std::function<double(double)>& derivative_activation_function,
+		                                  const std::function<float(float)>& derivative_activation_function,
 		                                  mat_arr* output_layer_error_rv) const override;
 	};
 }
