+++ annlib/net_init.h
@@ -18,7 +18,7 @@ namespace annlib
 	class gaussian_net_init : public net_init
 	{
 	private:
-		std::normal_distribution<double> distribution;
+		std::normal_distribution<float> distribution;
 		std::default_random_engine rng;
 	public:
 		gaussian_net_init();
